requirements.txt
# API
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# LangChain
langchain==0.3.0
langchain-community==0.3.0
langchain-groq==0.2.0
langchain-huggingface==0.1.0

# pgvector
pgvector==0.3.5
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.35
psycopg2-binary==2.9.9

# Loaders
pypdf==5.0.1
docx2txt==0.8

# Embeddings + reranker local CPU
sentence-transformers==3.1.1

# Hybrid search BM25
rank-bm25==0.2.2

# Config
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Streamlit
streamlit==1.39.0
requests==2.32.3

# Tests
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# Dev
ruff==0.6.9